Python/Lab/_gDrive/requirements.txt
httplib2==0.18.1
google_auth_oauthlib==0.4.1
oauth2client==4.1.3
PyDrive==1.3.1
google_api_python_client==1.12.4
protobuf==3.13.0
python-dotenv==0.14.0